update(create-adapter): Fix bug, improve maintainability & testability (v2)
- Fixed `transformInput` to incorrectly transforms string arrays
- Added `supportsArrays` and `supportsNumbers` in the adapter config. I realized I needed this when developing the Redis adapter.
- Made create adapter config’s `supportNumber`, `supportString`, etc all required, rather than using defaults which could be incorrect, this requires the adapter maintainer to think about all of the types their database supports.
- Moved a lot of the logic in the create-adapter out to it’s separate files to allow the createAdapter code to be less dense and drastically make it easier to read & maintain.
  - Also added tests for some of it's logic - such as `transformInput`
- Improved/updated the create-a-database-adapter docs

My previous [PR](https://github.com/better-auth/better-auth/pull/3265) I messed up the rebase so it was just easier to open a new PR (this one)
