+++ packages/better-auth/src/adapters/create-adapter/types.ts
@@ -61,18 +61,30 @@ export interface AdapterConfig {
 	adapterId: string;
 	/**
 	 * If the database supports numeric ids, set this to `true`.
+	 * For example, MongoDB doesn't support numeric ids.
+	 * Where as PostgreSQL does via `serial` columns.
 	 *
 	 * @default true
 	 */
-	supportsNumericIds?: boolean;
+	supportsNumericIds: boolean;
 	/**
 	 * If the database doesn't support JSON columns, set this to `false`.
 	 *
-	 * We will handle the translation between using `JSON` columns, and saving `string`s to the database.
+	 * Note: We do not fully support this yet, so
+	 * we will handle the translation between using `JSON` columns, and saving `string`s to the database.
 	 *
 	 * @default false
 	 */
-	supportsJSON?: boolean;
+	supportsJSON: boolean;
+	/**
+	 * If the database doesn't support JSONB columns, set this to `false`.
+	 *
+	 * Note: We do not fully support this yet, so
+	 * we will handle the translation between using `JSONB` columns, and saving `string`s to the database.
+	 *
+	 * @default false
+	 */
+	supportsJSONB: boolean;
 	/**
 	 * If the database doesn't support dates, set this to `false`.
 	 *
@@ -80,7 +92,7 @@ export interface AdapterConfig {
 	 *
 	 * @default true
 	 */
-	supportsDates?: boolean;
+	supportsDates: boolean;
 	/**
 	 * If the database doesn't support booleans, set this to `false`.
 	 *
@@ -88,7 +100,7 @@ export interface AdapterConfig {
 	 *
 	 * @default true
 	 */
-	supportsBooleans?: boolean;
+	supportsBooleans: boolean;
 	/**
 	 * Disable id generation for the `create` method.
 	 *
@@ -96,7 +108,23 @@ export interface AdapterConfig {
 	 *
 	 * @default false
 	 */
-	disableIdGeneration?: boolean;
+	disableIdGeneration: boolean;
+	/**
+	 * If the database doesn't support numbers, set this to `false`.
+	 *
+	 * We will handle the translation between using `number`s, and saving `string`s to the database.
+	 *
+	 * @default true
+	 */
+	supportsNumbers: boolean;
+	/**
+	 * If the database doesn't support arrays, set this to `false`.
+	 *
+	 * We will handle the translation between using `array`s, and saving `string`s to the database.
+	 *
+	 * @default true
+	 */
+	supportsArrays: boolean;
 	/**
 	 * Map the keys of the input data.
 	 *
@@ -381,3 +409,11 @@ export type AdapterTestDebugLogs = {
 	resetDebugLogs: () => void;
 	printDebugLogs: () => void;
 };
+
+export type IdField = ({
+	customModelName,
+	forceAllowId,
+}: {
+	customModelName?: string;
+	forceAllowId?: boolean;
+}) => FieldAttribute;
